minor adjustments

+++ src/com/polsl/model/PathMaker.java
@@ -66,7 +66,7 @@ public class PathMaker {
 
     public Path createPath(LinkedList<City> cities){
 
-        Integer distance=0;
+        Double distance=0.0;
         for (int i =0 ;i<cities.size()-1;i++) {
             distance+= cities.get(i).distanceTo(cities.get(i+1));
         }
